+++ outputs.tf
@@ -4,3 +4,7 @@ output "domain_name" {
 output "hosted_zone_id" {
   value = "${aws_cloudfront_distribution.distribution.hosted_zone_id}"
 }
+
+output "distribution_id" {
+  value = "${aws_cloudfront_distribution.distribution.id}"
+}