+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## backwardcomp/0.11
+ENHANCEMENTS:
+* Add output for Cloud Front Distribution Id
+* Add Origin Access Identity as variable
+
 ## 0.0.2 (Unreleased)
 ENHANCEMENTS: